test: renderPropsPipe
