+++ libs/frontend/modules/builder/src/renderer/pipes/types.ts
@@ -16,6 +16,12 @@ export interface RenderContext {
   /** The rendered tree */
   tree: ElementTree
 
+  render: (
+    element: IElement,
+    context: RenderContext,
+    props: RenderProps,
+  ) => RenderOutput
+
   /** Extra props passed to all element. They override the common props, but props from the node instance override the extraProps */
   extraProps?: RenderProps
